Raplaced any with {} on state definition for stateless objects, fixed typo, moved ConditionSelectorContainer to new file, added events for test page logging
