+++ src/components/ConditionSelector/index.ts
@@ -4,3 +4,4 @@ export * from './ConditionGroup';
 export * from './ConditionGroup.Props';
 export * from './ConditionSelector';
 export * from './ConditionSelector.Props';
+export * from './ConditionSelectorContainer';
